move iotda under noncore
